+++ app/controllers/application_controller.rb
@@ -2,6 +2,7 @@
 
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
+
   before_action :set_timezone
 
   before_action :assign_navbar_content