+++ app/models/transaction.rb
@@ -30,13 +30,16 @@ class Transaction < ApplicationRecord
 
   scope :with_balance, -> { includes(:transaction_balance).references(:transaction_balance) }
   scope :desc, -> { order('pending DESC, trx_date DESC, id DESC') }
-
-  scope :search, lambda { |query|
-    query = sanitize_sql_like(query)
-    where(arel_table[:description]
-            .lower
-            .matches("%#{query.downcase}%"))
-  }
+  scope :current_day, -> { where(created_at: Time.current.all_day) }
+  scope :pending, -> { where(pending: true) }
+  
+  def self.search(description)
+    if description
+      where('description ILIKE ?', "%#{description}%")
+    else
+      all
+    end
+  end
 
   # Determine the transaction_type for existing records based on amount
   def transaction_type